The game "Set of numbers" with a computer
Enter your own version of the set of numbers (from the above) from the keyboard and compare it with the computer's set of numbers, which is randomly selected from the 5 sets above. If the sum of the digits of your set of numbers is greater than the sum of the digits of the computer's set of numbers, then you have won (and vice versa). In case of identical sums of digits, it is a draw.